requirements_fr.txt
click==7.1.2
cycler==0.10.0
decorator==4.4.2
dlib==19.19.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
google-images-download==2.8.0
imagecodecs==2020.2.18
imageio==2.8.0
img2pdf==0.3.3
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.3
networkx==2.4
numpy==1.18.1
opencv-contrib-python==4.2.0.32
opencv-python==4.2.0.32
pdf2image==1.12.1
pdf2jpg==1.0
Pillow==7.0.0
pyparsing==2.4.6
pytesseract==0.3.2
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scikit-learn==0.20.3
scipy==1.4.1
selenium==3.141.0
six==1.14.0
threadpoolctl==2.0.0
tifffile==2020.5.11
urllib3==1.25.9
Werkzeug==1.0.1
